refactor: 예외를 token manager에서 던짐
