+++ api/src/main/kotlin/org/monsing/auth/AuthInterceptor.kt
@@ -2,24 +2,28 @@ package org.monsing.auth
 
 import jakarta.servlet.http.HttpServletRequest
 import jakarta.servlet.http.HttpServletResponse
+import org.monsing.jwt.TokenManager
 import org.springframework.stereotype.Component
 import org.springframework.web.method.HandlerMethod
 import org.springframework.web.servlet.HandlerInterceptor
-import javax.security.auth.login.CredentialException
 
 @Component
-class AuthInterceptor(private val tokenInterceptor: TokenInterceptor) : HandlerInterceptor {
+class AuthInterceptor(
+    private val tokenManager: TokenManager,
+    private val authContext: AuthContext
+) : HandlerInterceptor {
 
     override fun preHandle(request: HttpServletRequest, response: HttpServletResponse, handler: Any): Boolean {
         handler as HandlerMethod
 
         if (handler.hasMethodAnnotation(Auth::class.java)) {
-            val header = request.getHeader("Authorization") ?: throw CredentialException("No Authorization header")
-            require(header.startsWith("Bearer ")) {
-                "Invalid Authorization Header"
-            }
+            val token = request.getHeader("Authorization")
+                .takeIf { it.startsWith("Bearer ") }
+                ?.apply { removePrefix("Bearer ") }
+                ?: throw IllegalArgumentException("잘못된 인증 요청")
 
-            return tokenInterceptor.preHandle(request, response, handler)
+            val payload = tokenManager.getPayLoad(token)
+            authContext.payload = payload
         }
 
         return true
