--- api/src/main/kotlin/org/monsing/auth/ExpiredTokenException.kt
@@ -1,3 +1,0 @@
-package org.monsing.auth
-
-class ExpiredTokenException(s: String) : RuntimeException()
